+++ src/main/java/com/gabia/gyebalja/controller/ImageApiController.java
@@ -1,6 +1,6 @@
 package com.gabia.gyebalja.controller;
 
-import com.gabia.gyebalja.service.BoardImgService;
+import com.gabia.gyebalja.service.ImageService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -11,14 +11,22 @@ import java.io.IOException;
 
 @RequiredArgsConstructor
 @RestController
-public class BoardImgApiController {
+public class ImageApiController {
 
-    private final BoardImgService boardImgService;
+    private final ImageService imageService;
 
     /** 등록 - boardImg 한 건 (이미지 등록) */
     @PostMapping("/api/v1/boardImgs")
     public String postOneBoardImg(@RequestParam("image") MultipartFile image) throws IOException {
-        String response = boardImgService.postOneBoardImg(image);
+        String response = imageService.postOneBoardImg(image);
+
+        return response;
+    }
+
+    /** 등록 - userImg 한 건 (이미지 등록) */
+    @PostMapping("/api/v1/userImgs")
+    public String postOneUserImg(@RequestParam("image") MultipartFile image) throws IOException {
+        String response = imageService.postOneUserImg(image);
 
         return response;
     }
